refactor(sdk): cluster private credential collaborators under credentials/_internal
TransportKeyPairVault, PermissionStore, permissions.ts, and
storage-keys.ts each have zero consumers outside the CredentialService
cluster (verified by grep) — they're CredentialService's private
implementation, not peers in the credentials/ folder listing. types.ts,
schemas.ts, and utils.ts stay at the top level: each has a genuine
external consumer (config/build.ts, services/decryption-service.ts, the
credentials/index.ts barrel).

No public API impact — none of the moved symbols were ever exported.

SDK-196
